# openapi.yaml
openapi: 3.1.0
info:
  title: CivicLens API
  version: 1.0.0
  description: >
    AI-assisted public-service navigation API. Eligibility determinations are
    always produced by the deterministic rule engine (see
    docs/ai/eligibility-engine.md); the assistant endpoints are
    retrieval-grounded and always cite sources.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /schemes:
    get:
      summary: Browse/search schemes
      parameters:
        - in: query
          name: q
          schema: { type: string }
          description: Full-text / semantic search query
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: scope
          schema: { type: string, enum: [central, state] }
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        "200":
          description: Paginated scheme list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SchemePage' }
  /schemes/{scheme_id}:
    get:
      summary: Get scheme detail (current published version)
      parameters:
        - $ref: '#/components/parameters/SchemeIdParam'
      responses:
        "200":
          description: Scheme detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SchemeDetail' }
        "404": { $ref: '#/components/responses/NotFound' }

  # ---------- Eligibility ----------
...
  /admin/schemes/{scheme_id}/versions:
    post:
      summary: Create a new draft scheme_version
      parameters:
        - $ref: '#/components/parameters/SchemeIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SchemeVersionInput' }
      responses:
        "201":
          description: Draft version created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SchemeVersion' }
  /admin/scheme-versions/{scheme_version_id}/publish:
    post:
      summary: Publish a reviewed scheme_version (requires second reviewer)
      parameters:
        - name: scheme_version_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Published }
        "403": { description: Publisher must differ from author (four-eyes rule) }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSizeParam:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    SchemeIdParam:
      name: scheme_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string }
            message: { type: string }
            request_id: { type: string }
            field_errors:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
...
    SchemePage:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/SchemeSummary' }
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
    SchemeSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        canonical_name: { type: string }
        category: { type: string }
        scope: { type: string, enum: [central, state] }
        benefits_summary: { type: string }
    SchemeDetail:
      allOf:
        - $ref: '#/components/schemas/SchemeSummary'
        - type: object
          properties:
            administering_dept: { type: string }
            document_requirements:
              type: array
              items: { $ref: '#/components/schemas/DocumentRequirement' }
            last_verified_at: { type: string, format: date-time }
            scheme_version_id: { type: string, format: uuid }
    DocumentRequirement:
      type: object
      properties:
        document_type: { type: string }
        is_mandatory: { type: boolean }
        notes: { type: string, nullable: true }
    SchemeVersionInput:
      type: object
      required: [benefits_summary, effective_from]
      properties:
        benefits_summary: { type: string }
        effective_from: { type: string, format: date }
        knowledge_source_id: { type: string, format: uuid }
    SchemeVersion:
      type: object
      properties:
        id: { type: string, format: uuid }
        scheme_id: { type: string, format: uuid }
        version_no: { type: integer }
        status: { type: string, enum: [draft, in_review, published, superseded, archived] }
...